requirements.txt
fastapi
uvicorn
pydantic
google-genai
taibun
dotenv
pillow
python-multipart